+++ src/Player.java
@@ -18,13 +18,13 @@ public class Player extends Cell {
 	public void moveAndDraw(Graphics2D g2d)
 	{
 		g2d.setColor(new Color(0,0,0,0)); //paints w/ background color to erase
-		g2d.fillOval((int) super.getLocation().getX() - super.getRadius(), (int) super.getLocation().getY() - super.getRadius(), super.getRadius() * 2, super.getRadius() * 2); //erases current oval
+		g2d.fillOval((int) (super.getLocation().getX() - super.getRadius()), (int) (super.getLocation().getY() - super.getRadius()), (int) super.getRadius() * 2, (int) super.getRadius() * 2); //erases current oval
 		
 		super.setLocation(super.getNewLocation());
 		super.setNewLocation(super.getLocation());
 		
 		g2d.setColor(super.getColor());
-		g2d.fillOval((int) super.getLocation().getX() - super.getRadius(), (int) super.getLocation().getY() - super.getRadius(), super.getRadius() * 2, super.getRadius() * 2); //draws new oval
+		g2d.fillOval((int) (super.getLocation().getX() - super.getRadius()), (int) (super.getLocation().getY() - super.getRadius()), (int) super.getRadius() * 2, (int) super.getRadius() * 2); //draws new oval
 		
 		String name = new String("Amorba!!!");
 		AffineTransform affinetransform = new AffineTransform();
